Update: Move where cardinality sizes are exported from
