+++ packages/data/addon/utils/enums/cardinality-sizes.js
@@ -1,0 +1,5 @@
+/**
+ * Copyright 2020, Yahoo Holdings Inc.
+ * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
+ */
+export default ['SMALL', 'MEDIUM', 'LARGE'];
