+++ packages/data/app/utils/enums/cardinality-sizes.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-data/utils/enums/cardinality-sizes';
